Crate gbdt

Modules

binary_tree

This module implements a simple binary tree.

config

This module implements the config for gradient boosting.

decision_tree

This module implements a decision tree from the simple binary tree gbdt::binary_tree.

fitness

This module implements some math functions used for gradient boosting process.

gradient_boost

This module implements the process of gradient boosting decision tree algorithm. This module depends on the following module:

input

This module implements the data loader.