Module gbdt::fitness

This module implements some math functions used for gradient boosting process.

Functions

AUC
AUC (Area Under the Curve) calculation for first n element in data vector. See wikipedia for detailed algorithm.

MAE
MAE (Mean Absolute Error) calculation for first n element in data vector. See wikipedia for detail for detailed algorithm.

RMSE
RMSE (Root-Mean-Square deviation) calculation for first n element in data vector. See wikipedia for detailed algorithm.

almost_equal
Comparing two number with default floating error threshold.

almost_equal_thrs
Comparing two number with a costomized floating error threshold.

average
Return the weighted target average for first n data in data vector.

label_average
Return the weighted label average for first n data in data vector.

lad_loss
LAD loss function.

lad_loss_gradient
LAD gradient calculation.

logit
Logistic value function.

logit_loss
Negative binomial log-likelyhood loss function.

logit_loss_gradient
Log-likelyhood gradient calculation.

same
Return whether the first n data in data vector have same target values.

weighted_label_median
Return the weighted label median for first n data in data vector.

weighted_residual_median
Return the weighted residual median for first n data in data vector.